chore: move the rest of breakpoint algorithms to `@makeswift/controls`
