+++ .changeset/good-worms-crash.md
@@ -1,0 +1,6 @@
+---
+"@makeswift/controls": patch
+"@makeswift/runtime": patch
+---
+
+chore: move the rest of breakpoint algorithms to the `@makeswift/controls` package
