+++ packages/controls/package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "color": "^3.1.3",
     "css-box-model": "^1.2.1",
+    "csstype": "^2.6.5",
     "slate": "0.91.4",
     "ts-pattern": "^5.0.8",
     "uuid": "^9.0.0",
