+++ packages/runtime/src/components/builtin/Carousel/Carousel.tsx
@@ -13,6 +13,8 @@ import { motion, useAnimation } from 'framer-motion'
 import { useGesture } from '@use-gesture/react'
 import { wrap } from '@popmotion/popcorn'
 
+import { getBaseBreakpoint } from '@makeswift/controls'
+
 import { colorToString } from '../../utils/colorToString'
 import { type ResponsiveColor } from '../../utils/types'
 import { useMediaQuery } from '../../hooks'
@@ -21,7 +23,6 @@ import Image from '../Image'
 import { useStyle } from '../../../runtimes/react/use-style'
 import { cx } from '@emotion/css'
 import { useResponsiveStyle } from '../../utils/responsive-style'
-import { getBaseBreakpoint } from '../../../state/modules/breakpoints'
 import { useBreakpoints } from '../../../runtimes/react/hooks/use-breakpoints'
 import {
   type ResponsiveGapData,
