+++ packages/runtime/src/core/index.ts
@@ -6,6 +6,7 @@ export {
   type Data,
   type DataType,
   type DeviceOverride,
+  type FallbackStrategy,
   type ValueType,
   type ResolvedValueType,
   type ResponsiveValue,
@@ -16,8 +17,12 @@ export {
   ControlInstance,
   DefaultControlInstance,
   ShapeV2Control,
+  getBaseBreakpoint,
+  getBreakpoint,
   findBreakpointOverride,
+  findNextFallback,
   mergeOrCoalesceFallbacks,
   mergeResponsiveValues,
   shallowMergeFallbacks,
+  getViewportStyle,
 } from '@makeswift/controls'
